+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 PaperTrail::Rails::Engine.eager_load!
 ```
 
+  - [#440](https://github.com/airblade/paper_trail/pull/440) - `versions` association should clear/reload after a transaction rollback.
   - [#438](https://github.com/airblade/paper_trail/issues/438) - `Model.paper_trail_enabled_for_model?` should return `false` if
     `has_paper_trail` has not been declared on the class.
   - [#427](https://github.com/airblade/paper_trail/pull/427) - Fix `reify` method in context of model where a column has been removed.